Add topics with extension point information
These topics contains reference information that was removed from the plug-ins topics, as they were refactored as task topics. I've edited the descriptions of the extension points for clarity and consistency. I know that we now are programmatically-generating a topic that lists the extension points, so I'm assuming that these topics are temporary -- but we don't want to lose the descriptions of the extension points, which are more content rich than the descriptions in the generated topic.