Update processing flow chart (Fixes #54)
This commit replaces the old image with a new PNG placeholder and includes the original source file and an SVG version that can be swapped in later to improve output quality if upgrading FOP resolves the font rendering issues in the current PDF output.

The chart was drawn in OmniGraffle 6.6.1, but the SVG output can also be edited in Inkscape (inkscape.org).

Signed-off-by: Roger Sheen <roger@infotexture.net>