Tweaking technically innacurate (obsolete) info
The docs currently use HTML5 as an example that extends another transform, but as of 2.4 this is no longer true. I've tweaked the language to point out that this remains a good example, but technically applies to earlier versions of that plugin.

Potential longer term alternatives: create a full example, use XHTML (which is much longer and sort of out of date), or HTML Help (which is really short but REALLY out of date), or just use an abbreviated version of another plugin.

Signed-off-by: Robert D Anderson <robander@us.ibm.com>