Restore autogenerated properties template
When the docs Ant build was replaced with the new Gradle build for #178, the `generate-properties-file` target created for 2.3 with #60 was inadvertently left behind.

Since the distribution build for the core repo was updated to use the docs Gradle build for 3.1 per dita-ot/dita-ot#2972, this file is no longer generated when the docs build runs.

This commit ports the missing target from the old Ant build to the current Gradle build file to ensure that the template is once again available in the `docsrc/samples/properties/` directory of distribution builds.

Signed-off-by: Roger Sheen <roger@infotexture.net>