Align command-line examples with `dita` in $PATH
The command-line samples we provide to illustrate how to do various things with the `dita` command currently use a variety of paths to refer to the `dita` command depending on the assumed current directory.

- `dita-ot-dir/bin/dita …`
- `bin/dita …`
- `dita …`

These changes adjust the installation instructions to explicitly include the addition of the DITA-OT `bin` directory to $PATH, and simplify the sample commands to begin with `dita …`.

Signed-off-by: Roger Sheen <roger@infotexture.net>