Move remaining properties to Gradle build script
The properties files were originally created in 9cceb75 to allow the same properties to be used in both Ant and Gradle builds. The legacy Ant build was removed ac79e97 for #178, so there's no longer any need to maintain properties files that are separate from the build script.

Since ffd660f, properties that require absolute paths are set in the Gradle build to anchor these paths to the Gradle project directory, which corresponds to the `docsrc/` folder of the DITA-OT distribution package, or the root folder of the docs repository when cloned to a standalone location.

This supports location-independent builds in a more reliable fashion than the previous mechanism that relied on the Ant `${basedir}` property, which is interpreted differently in recent Gradle and JDK versions.

Signed-off-by: Roger Sheen <roger@infotexture.net>